+++ apps/playground/src/app/components/forms/field-group/field-group.component.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { Component, inject } from '@angular/core';
 import {
   FormBuilder,
@@ -6,17 +7,33 @@ import {
   FormsModule,
   ReactiveFormsModule,
 } from '@angular/forms';
-import { SkyFieldGroupModule, SkyInputBoxModule } from '@skyux/forms';
+import { SkyIdModule } from '@skyux/core';
+import {
+  SkyFieldGroupModule,
+  SkyInputBoxModule,
+  SkyRadioModule,
+  SkyToggleSwitchModule,
+} from '@skyux/forms';
+
+interface Item {
+  icon: string;
+  label: string;
+  name: string;
+}
 
 @Component({
   selector: 'app-field-group',
   templateUrl: './field-group.component.html',
   standalone: true,
   imports: [
+    CommonModule,
     FormsModule,
     ReactiveFormsModule,
     SkyFieldGroupModule,
+    SkyIdModule,
     SkyInputBoxModule,
+    SkyRadioModule,
+    SkyToggleSwitchModule,
   ],
 })
 export class FieldGroupComponent {
@@ -24,9 +41,18 @@ export class FieldGroupComponent {
 
   #formBuilder = inject(FormBuilder);
 
+  protected views: Item[] = [
+    { icon: 'table', label: 'Table', name: 'table' },
+    { icon: 'list', label: 'List', name: 'list' },
+    { icon: 'map-marker', label: 'Map', name: 'map' },
+  ];
+
   constructor() {
     this.formGroup = this.#formBuilder.group({
       name: new FormControl(undefined),
+      hometown: new FormControl(undefined),
+      toggle: new FormControl(false),
+      radio: this.views[0].name,
     });
   }
 }