+++ libs/components/forms/src/lib/modules/checkbox/checkbox.component.ts
@@ -3,9 +3,11 @@ import {
   Component,
   ElementRef,
   EventEmitter,
+  HostBinding,
   Input,
   OnInit,
   Output,
+  TemplateRef,
   ViewChild,
   booleanAttribute,
   inject,
@@ -16,6 +18,7 @@ import { SkyFormsUtility, SkyIdService, SkyLogService } from '@skyux/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 
 import { SKY_FORM_ERRORS_ENABLED } from '../form-error/form-errors-enabled-token';
+import { SkyFormFieldLabelTextRequiredService } from '../shared/form-field-label-text-required.service';
 
 import { SkyCheckboxChange } from './checkbox-change';
 
@@ -131,6 +134,23 @@ export class SkyCheckboxComponent implements ControlValueAccessor, OnInit {
   public get name(): string {
     return this.#_name;
   }
+
+  /**
+   * The content of the help popover. When specified along with `labelText`, a [help inline](https://developer.blackbaud.com/skyux/components/help-inline)
+   * button is added to the checkbox label. The help inline button displays a [popover](https://developer.blackbaud.com/skyux/components/popover)
+   * when clicked using the specified content and optional title.
+   * @preview
+   */
+  @Input()
+  public helpPopoverContent: string | TemplateRef<unknown> | undefined;
+
+  /**
+   * The title of the help popover. This property only applies when `helpPopoverContent` is
+   * also specified.
+   * @preview
+   */
+  @Input()
+  public helpPopoverTitle: string | undefined;
 
   /**
    * Fires when the selected value changes.
@@ -285,6 +305,9 @@ export class SkyCheckboxComponent implements ControlValueAccessor, OnInit {
     return this.#_inputEl;
   }
 
+  @HostBinding('style.display')
+  public display: string | undefined;
+
   protected inputId = '';
 
   #checkedChange: BehaviorSubject<boolean>;
@@ -313,6 +336,11 @@ export class SkyCheckboxComponent implements ControlValueAccessor, OnInit {
     optional: true,
     self: true,
   });
+
+  readonly #labelTextRequired = inject(SkyFormFieldLabelTextRequiredService, {
+    optional: true,
+  });
+
   protected readonly errorId = this.#idSvc.generateId();
 
   constructor() {
@@ -338,6 +366,11 @@ export class SkyCheckboxComponent implements ControlValueAccessor, OnInit {
       this.required =
         this.required || SkyFormsUtility.hasRequiredValidation(this.ngControl);
     }
+
+    if (this.#labelTextRequired && !this.labelText) {
+      this.display = 'none';
+    }
+    this.#labelTextRequired?.validateLabelText(this.labelText);
   }
 
   /**